[System Page]Step 2: Added rest of the system page functions in a separate file and imported them. Status: Everything works!
